+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - An error message is displayed when the report task to redmine fails
 
+### Added
+
+- Show task preview after modify it
+
 ## [0.1.2] - 2023-06-03
 
 ### Added